requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
websockets==16.0
aiofiles==25.1.0
python-multipart==0.0.21
jinja2==3.1.6
pydantic==2.12.5
click==8.3.1
rich==14.2.0
plyer==2.1.0
psutil==7.2.1
requests==2.32.5
httpx==0.28.1
aiohttp==3.13.3
cryptography==46.0.3
python-jose[cryptography]==3.5.0
apscheduler==3.11.2
aiosmtplib==5.1.0
playwright==1.50.0
pytest==9.0.2
pytest-asyncio==1.3.0
